luci-0.9: merge r5175

+++ libs/lmo/src/lmo_core.c
@@ -71,7 +71,7 @@ lmo_archive_t * lmo_open(const char *file)
 		goto cleanup;
 	}
 
-	if( lseek(in, (off_t)(-sizeof(uint32_t)), SEEK_END) == -1 )
+	if( lseek(in, -sizeof(uint32_t), SEEK_END) == -1 )
 	{
 		error("Can not seek to eof", 1);
 		goto cleanup;
